// src/database/common/enum/serialization-group.enum.ts
export enum SERIALIZE_GROUP {
  GROUP_BASE = 'GROUP_BASE',
  GROUP_ALL_BASE = 'GROUP_ALL_BASE',

  GROUP_ADMIN = 'GROUP_ADMIN',
  GROUP_ALL_ADMIN = 'GROUP_ALL_ADMIN',

  GROUP_ALL_ROLE = 'GROUP_ALL_ROLE',
  GROUP_ROLE = 'GROUP_ROLE',

  GROUP_ALL_USER = 'GROUP_ALL_USER',
  GROUP_USER_PROFILE = 'GROUP_USER_PROFILE',
  GROUP_USER = 'GROUP_USER',
}